+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.4.1 (unreleased)
 ...................
 * 404 for ``.well-known/*`` urls in  ``spa_static_handler``
+* allow google recaptcha test keys
 
 v0.4.0 (2019-04-09)
 ...................